Pull shared editor code out to a hook
